Name add

+++ src/main/java/com/miss/artificial_city/model/car/Car.java
@@ -46,7 +46,7 @@ public class Car {
     }
 
     private void moveNodes(Node node) {
-        for (int i = 0; i < currentSpeed; i++) {
+        for (int i = 0; i < currentSpeed.intValue(); i++) {
             node = node.getNeighbors().getRight();
         }
         setPresenceToNewNodes(node);
